requirements.txt
black==21.9b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
dataclasses==0.8
google-auth==2.3.1
idna==3.3
importlib-metadata==4.8.1
kubernetes==18.20.0
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
pick==1.0.0
platformdirs==2.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
PyYAML==6.0
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tomli==1.2.2
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
websocket-client==1.2.1
zipp==3.6.0
